Added SensorMotor class

+++ lib/WebConfigServer/WebConfigServer.h
@@ -86,7 +86,7 @@ public:
 
   struct Device {
     bool track_restart_counter;
-    int angle_accuracy;
+    float angle_accuracy;
     float servo_speed_ms;
     int loop_sensor_time_ms;
     int publish_time_ms;
